Split tests for each generator into separate files
